python_consumer/requirements.txt
cloudinary==1.44.0
opencv-python==4.11.0.86
pika==1.3.2
pydantic-settings==2.9.1
pymongo[srv]==4.13.1
pillow==11.2.1